requirements.txt
# Airflow dependencies
apache-airflow==2.7.2
apache-airflow-providers-sqlite==3.4.0

# Flask dependencies (compatible versions for Airflow)
Flask==2.2.5
Flask-Session==0.4.0

# dbt with SQLite
dbt-core==1.9.1
dbt-sqlite==1.9.1

# SQLite interaction
sqlalchemy==1.4.37
pandas==1.5.3

# General utility
python-dotenv==0.21.1
requests==2.28.2

# Optional: for CSV loading
pyyaml==6.0

# Connexion (for Airflow API handling)
connexion[swagger-ui]==2.14.2